backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
Pillow==10.4.0
django-allauth==0.63.6
gunicorn==21.2.0
whitenoise==6.6.0
channels==4.1.0
daphne==4.1.2